refactor(LottoMachine): move LottoMAchine from model to controller
