+++ src/main/kotlin/lotto/controller/Controller.kt
@@ -1,7 +1,5 @@
 package lotto.controller
 
-import lotto.model.LottoMachine
-
 fun main() {
     LottoMachine.start()
 }
