+++ src/main/kotlin/lotto/controller/LottoMachine.kt
@@ -1,10 +1,13 @@
-package lotto.model
+package lotto.controller
 
+import lotto.model.LottoNumber
+import lotto.model.ManualTicket
+import lotto.model.Ticket
 import lotto.service.AmountValidator
 import lotto.service.Constants
+import lotto.service.Rank
 import lotto.view.InputView
 import lotto.view.OutputView
-import lotto.service.Rank
 
 object LottoMachine {
     fun start() {
@@ -69,7 +72,7 @@ object LottoMachine {
     }
 
     private fun generateTicketNumbers(): List<LottoNumber> {
-        return LottoNumber.takeRandom(Constants.NUMBER_COUNT)
+        return LottoNumber.Companion.takeRandom(Constants.NUMBER_COUNT)
     }
 
     private val winStats =
@@ -93,10 +96,10 @@ object LottoMachine {
                 winStats[Rank.SECOND] = winStats[Rank.SECOND]!! + 1
             } else {
                 winStats[
-                    Rank.valueOf(
+                    Rank.Companion.valueOf(
                         match, false,
                     ),
-                ] = winStats[Rank.valueOf(match, false)]!! + 1
+                ] = winStats[Rank.Companion.valueOf(match, false)]!! + 1
             }
         }
     }
